% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AUC_BS.R
\name{calc_AUC_BS}
\alias{calc_AUC_BS}
\title{Calculate time-dependent AUC and BS}
\usage{
calc_AUC_BS(
  l.formula,
  s.formula,
  long.test,
  surv.test,
  Tstart,
  Tstop,
  summary.mean,
  seed = 123
)
}
\arguments{
\item{l.formula}{the formula for longitudinal model}

\item{s.formula}{the formula for survival model}

\item{long.test}{the longitudinal test set}

\item{Tstart}{starting time}

\item{Tstop}{stop time}

\item{summary.mean}{estimated parameters value (mean)}

\item{seed}{random seed (default = 123)}

\item{surv.dat}{the survival test set}
}
\value{
a Stan fit object
}
\description{
This function allows you to pass in longitudinal and survival
formula, with the longitudinal and survival test dataset, 
start time, stop time and estimated parameters, to calculate
time-dependent Area Under Curve for ROC, and Brier Score
}
