% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AUC_BS_RSF.R
\name{calc_AUC_BS_RSF}
\alias{calc_AUC_BS_RSF}
\title{Calculate time-dependent AUC and BS for RSF}
\usage{
calc_AUC_BS_RSF(surv.test2, Tstart, Tstop, res)
}
\arguments{
\item{Tstart}{starting time}

\item{Tstop}{stop time}

\item{res}{predicted result using RSF on test dataset}

\item{surv.dat2}{the survival test set}
}
\value{
a length-2 vector: estimated AUC and BS
}
\description{
This function allows you to pass in survival test dataset, 
start time, stop time and predicted RSF result on test dataset, 
to calculate time-dependent Area Under Curve for ROC, and Brier Score
}
