% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv.R
\name{create_cv}
\alias{create_cv}
\title{Create cross-validation data}
\usage{
create_cv(long.dat, surv.dat, ratio = 3, reps = 10, seed = 123)
}
\arguments{
\item{long.dat}{the dataset for longitudinal model}

\item{surv.dat}{the dataset for survival model}

\item{ratio}{ratios of subjects in training dataset and testing dataset (default=3)}

\item{reps}{how many replicates needed (default=10)}

\item{seed}{random seed (default = 123)}
}
\value{
a length-reps list, each element contains 
longitudianl and survival training and testing dataset
}
\description{
This function allows you to pass in longitudinal and survival
datasets with the cv ratio and replicates, 
to create cross-validation partitions (training data and testing data).
}
\examples{

create_cv(
jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)[[1]],
jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)[[2]])


}
