% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Joint modeling using Stan}
\usage{
fit_stan(
  l.formula,
  s.formula,
  long.dat,
  surv.dat,
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  seed = 123,
  progress = TRUE
)
}
\arguments{
\item{l.formula}{the formula for longitudinal model}

\item{s.formula}{the formula for survival model}

\item{long.dat}{the longitudinal dataset}

\item{surv.dat}{the survival dataset}

\item{n.iter}{total number of iterations in NUTS (default = 2000)}

\item{n.burnin}{number of burnins in NUTS (default = n.iter/2)}

\item{seed}{random seed (default = 123)}

\item{progress}{whether to show progress in sampling (default = TRUE)}
}
\value{
a Stan fit object
}
\description{
This function allows you to pass in longitudinal and survival
formula, with the longitudinal and survival dataset, 
to perform NUTS sampling using Stan
}
\examples{

fit_stan(CD4 ~ obstime, Surv(Time, death) ~ drugddI, 
jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)[[1]],
jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)[[2]])

}
