% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_filter.R
\name{jm_filter}
\alias{jm_filter}
\title{Create the longitudinal and survival datasets}
\usage{
jm_filter(l.formula, s.formula, dat, ID)
}
\arguments{
\item{l.formula}{the formula for longitudinal model}

\item{s.formula}{the formula for survival model}

\item{dat}{original dataset}

\item{ID}{ID variable (can be factor, numeric or character)}
}
\value{
a length-2 list: longitudinal and survival dataframe
}
\description{
This function allows you to pass in longitudinal and survival
formula, with the full dataset and ID variable, to create two
separate datasets: longitudinal dataset and survival dataset.
}
\examples{

jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)

}
