% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCEM.R
\name{MCEM}
\alias{MCEM}
\title{Joint modeling using MCEM}
\usage{
MCEM(
  l.formula,
  s.formula,
  long.dat,
  surv.dat,
  max.iter = 30,
  tol = 0.001,
  seed = 123
)
}
\arguments{
\item{l.formula}{the formula for longitudinal model}

\item{s.formula}{the formula for survival model}

\item{long.dat}{the longitudinal dataset}

\item{surv.dat}{the survival dataset}

\item{max.iter}{maximum iterations allowed (default = 30)}

\item{tol}{tolerance (default = 1e-3)}

\item{seed}{random seed (default = 123)}
}
\value{
a list of estimated parameters
}
\description{
This function allows you to pass in longitudinal and survival
formula, with the longitudinal and survival dataset, 
to perform EM algorithm using Metropolis Hastings random walk sampler
}
\examples{

MCEM(CD4 ~ obstime, Surv(Time, death) ~ drugddI, 
jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)[[1]],
jm_filter(CD4 ~ obstime, Surv(Time, death) ~ drug, JM::aids, JM::aids$patient)[[2]])

}
