% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.Bayesian.R
\name{sim.Bayesian}
\alias{sim.Bayesian}
\title{Simulation for Bayesian method}
\usage{
sim.Bayesian(
  sim.num = 10,
  n.iter = 2000,
  n.burnin = floor(n.iter/2),
  seed = 123,
  progress = TRUE
)
}
\arguments{
\item{sim.num}{number of simulations (default = 10)}

\item{n.iter}{total number of iterations in NUTS (default = 2000)}

\item{n.burnin}{number of burnins in NUTS (default = n.iter/2)}

\item{seed}{random seed (default = 123)}

\item{progress}{whether to show progress for sampling (default = TRUE)}
}
\value{
a data frame containing parameter name, true values of parameters, 
mean of estimated parameters, SD of estimated parameters, square root of 
mean of SE^2 and coverage probability.
}
\description{
This function allows you to evaluate the correctness of Bayesian method
using simulated data
}
\examples{

sim.Bayesian(sim.num = 2)

}
