% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.MCEM.R
\name{sim.MCEM}
\alias{sim.MCEM}
\title{Simulation for MCEM method}
\usage{
sim.MCEM(sim.num = 10, max.iter = 30, tol = 0.001, seed = 123, progress = TRUE)
}
\arguments{
\item{sim.num}{number of simulations (default = 10)}

\item{max.iter}{maximum iterations allowed (default = 30)}

\item{tol}{tolerance (default = 1e-3)}

\item{seed}{random seed (default = 123)}

\item{progress}{whether to show progress for each iteration (default = TRUE)}
}
\value{
a data frame containing parameter name, true values of parameters, 
mean of estimated parameters, SD of estimated parameters.
}
\description{
This function allows you to evaluate the correctness of MCEM method
using simulated data
}
\examples{

sim.MCEM(sim.num = 2)

}
