% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateDataset.R
\name{SimulateDataset}
\alias{SimulateDataset}
\title{Simulate the longitudinal and survival datasets}
\usage{
SimulateDataset(n.train = 500, n.test = 200, seed)
}
\arguments{
\item{n.train}{number of subjects in training dataset (default=500)}

\item{n.test}{number of subjects in testing dataset (default=200)}

\item{seed}{random seed}
}
\value{
a length-4 list: longitudinal training, longitudinal testing, 
survival training, survival testing.
}
\description{
This function allows you to pass in subjects in training dataset
and testing dataset to create four separate datasets: 
longitudinal training, longitudinal testing, 
survival training, survival testing.
}
\examples{

l = SimulateDataset(seed = 123)
head(l[[1]])

}
